Update text classification to new api
