+++ flash/text/classification/data.py
@@ -11,196 +11,37 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from functools import partial
-from typing import Any, Callable, Dict, List, Optional, Sequence, Union
+from typing import Any, Callable, Dict, List, Optional, Sequence, Type, Union
 
-import pandas as pd
-import torch
 from pandas.core.frame import DataFrame
 
-from flash.core.data.data_module import DataModule
 from flash.core.data.data_pipeline import DataPipelineState
-from flash.core.data.io.classification_input import ClassificationInput, ClassificationState
-from flash.core.data.io.input import DataKeys, InputFormat
-from flash.core.data.io.input_transform import InputTransform
-from flash.core.data.utilities.classification import TargetMode
+from flash.core.data.io.input_base import Input
+from flash.core.data.new_data_module import DataModule
 from flash.core.data.utilities.paths import PATH_TYPE
 from flash.core.integrations.labelstudio.input import _parse_labelstudio_arguments, LabelStudioTextClassificationInput
-from flash.core.integrations.transformers.states import TransformersBackboneState
-from flash.core.utilities.imports import _TEXT_AVAILABLE, requires
+from flash.core.integrations.transformers.transforms import TransformersInputTransform
+from flash.core.utilities.imports import _TEXT_AVAILABLE
 from flash.core.utilities.stages import RunningStage
-from flash.text.input import TextDeserializer
+from flash.text.classification.input import (
+    TextClassificationCSVInput,
+    TextClassificationDataFrameInput,
+    TextClassificationInput,
+    TextClassificationJSONInput,
+    TextClassificationListInput,
+    TextClassificationParquetInput,
+)
 
 if _TEXT_AVAILABLE:
-    from datasets import Dataset, load_dataset
+    from datasets import Dataset
 else:
     Dataset = object
-
-
-class TextClassificationInput(ClassificationInput):
-    @staticmethod
-    def _resolve_target(target_keys: Union[str, List[str]], element: Dict[str, Any]) -> Dict[str, Any]:
-        if not isinstance(target_keys, List):
-            element[DataKeys.TARGET] = element.pop(target_keys)
-        else:
-            element[DataKeys.TARGET] = [element[target_key] for target_key in target_keys]
-        return element
-
-    @requires("text")
-    def load_data(
-        self,
-        hf_dataset: Dataset,
-        input_key: str,
-        target_keys: Optional[Union[str, List[str]]] = None,
-        max_length: int = 128,
-    ) -> Dataset:
-        """Loads data into HuggingFace datasets.Dataset."""
-        self.max_length = max_length
-
-        if not self.predicting:
-            hf_dataset = hf_dataset.map(partial(self._resolve_target, target_keys))
-            targets = hf_dataset.to_dict()[DataKeys.TARGET]
-            self.load_target_metadata(targets)
-
-            # If we had binary multi-class targets then we also know the labels (column names)
-            if self.target_mode is TargetMode.MULTI_BINARY and isinstance(target_keys, List):
-                classification_state = self.get_state(ClassificationState)
-                self.set_state(ClassificationState(target_keys, classification_state.num_classes))
-
-        # remove extra columns
-        extra_columns = set(hf_dataset.column_names) - {input_key, DataKeys.TARGET}
-        hf_dataset = hf_dataset.remove_columns(extra_columns)
-
-        if input_key != DataKeys.INPUT:
-            hf_dataset = hf_dataset.rename_column(input_key, DataKeys.INPUT)
-
-        return hf_dataset
-
-    def load_sample(self, sample: Dict[str, Any]) -> Any:
-        tokenized_sample = self.get_state(TransformersBackboneState).tokenizer(
-            sample[DataKeys.INPUT], max_length=self.max_length, truncation=True, padding="max_length"
-        )
-        tokenized_sample = tokenized_sample.data
-        if DataKeys.TARGET in sample:
-            tokenized_sample[DataKeys.TARGET] = self.format_target(sample[DataKeys.TARGET])
-        return tokenized_sample
-
-
-class TextClassificationCSVInput(TextClassificationInput):
-    @requires("text")
-    def load_data(
-        self,
-        csv_file: PATH_TYPE,
-        input_key: str,
-        target_keys: Optional[Union[str, List[str]]] = None,
-        max_length: int = 128,
-    ) -> Dataset:
-        dataset_dict = load_dataset("csv", data_files={"data": str(csv_file)})
-        return super().load_data(dataset_dict["data"], input_key, target_keys, max_length)
-
-
-class TextClassificationJSONInput(TextClassificationInput):
-    @requires("text")
-    def load_data(
-        self,
-        json_file: PATH_TYPE,
-        field: str,
-        input_key: str,
-        target_keys: Optional[Union[str, List[str]]] = None,
-        max_length: int = 128,
-    ) -> Dataset:
-        dataset_dict = load_dataset("json", data_files={"data": str(json_file)}, field=field)
-        return super().load_data(dataset_dict["data"], input_key, target_keys, max_length)
-
-
-class TextClassificationDataFrameInput(TextClassificationInput):
-    @requires("text")
-    def load_data(
-        self,
-        data_frame: pd.DataFrame,
-        input_key: str,
-        target_keys: Optional[Union[str, List[str]]] = None,
-        max_length: int = 128,
-    ) -> Dataset:
-        return super().load_data(Dataset.from_pandas(data_frame), input_key, target_keys, max_length)
-
-
-class TextClassificationParquetInput(TextClassificationInput):
-    @requires("text")
-    def load_data(
-        self,
-        parquet_file: PATH_TYPE,
-        input_key: str,
-        target_keys: Optional[Union[str, List[str]]] = None,
-        max_length: int = 128,
-    ) -> Dataset:
-        return super().load_data(Dataset.from_parquet(str(parquet_file)), input_key, target_keys, max_length)
-
-
-class TextClassificationListInput(TextClassificationInput):
-    @requires("text")
-    def load_data(
-        self,
-        inputs: List[str],
-        targets: Optional[List[Any]] = None,
-        max_length: int = 128,
-    ) -> Dataset:
-        if targets is not None:
-            hf_dataset = Dataset.from_dict({DataKeys.INPUT: inputs, DataKeys.TARGET: targets})
-        else:
-            hf_dataset = Dataset.from_dict({DataKeys.INPUT: inputs})
-        return super().load_data(hf_dataset, DataKeys.INPUT, DataKeys.TARGET, max_length)
-
-
-class TextClassificationInputTransform(InputTransform):
-    @requires("text")
-    def __init__(
-        self,
-        train_transform: Optional[Dict[str, Callable]] = None,
-        val_transform: Optional[Dict[str, Callable]] = None,
-        test_transform: Optional[Dict[str, Callable]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
-        max_length: int = 128,
-    ):
-        self.max_length = max_length
-
-        super().__init__(
-            train_transform=train_transform,
-            val_transform=val_transform,
-            test_transform=test_transform,
-            predict_transform=predict_transform,
-            inputs={
-                InputFormat.CSV: TextClassificationCSVInput,
-                InputFormat.JSON: TextClassificationJSONInput,
-                InputFormat.PARQUET: TextClassificationParquetInput,
-                InputFormat.HUGGINGFACE_DATASET: TextClassificationInput,
-                InputFormat.DATAFRAME: TextClassificationDataFrameInput,
-                InputFormat.LISTS: TextClassificationListInput,
-            },
-            default_input=InputFormat.LISTS,
-            deserializer=TextDeserializer(max_length),
-        )
-
-    def get_state_dict(self) -> Dict[str, Any]:
-        return {
-            **self.transforms,
-            "max_length": self.max_length,
-        }
-
-    @classmethod
-    def load_state_dict(cls, state_dict: Dict[str, Any], strict: bool):
-        return cls(**state_dict)
-
-    def per_sample_transform(self, sample: Dict[str, Any]) -> Dict[str, Any]:
-        for key in sample:
-            sample[key] = torch.as_tensor(sample[key])
-        return sample
 
 
 class TextClassificationData(DataModule):
     """Data Module for text classification tasks."""
 
-    input_transform_cls = TextClassificationInputTransform
+    input_transform_cls = TransformersInputTransform
 
     @classmethod
     def from_csv(
@@ -211,10 +52,12 @@ class TextClassificationData(DataModule):
         val_file: Optional[PATH_TYPE] = None,
         test_file: Optional[PATH_TYPE] = None,
         predict_file: Optional[PATH_TYPE] = None,
-        train_transform: Optional[Dict[str, Callable]] = None,
-        val_transform: Optional[Dict[str, Callable]] = None,
-        test_transform: Optional[Dict[str, Callable]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
+        train_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        val_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        test_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        predict_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        input_cls: Type[Input] = TextClassificationCSVInput,
+        transform_kwargs: Optional[Dict] = None,
         max_length: int = 128,
         **data_module_kwargs: Any,
     ) -> "TextClassificationData":
@@ -242,25 +85,20 @@ class TextClassificationData(DataModule):
             The constructed data module.
         """
 
-        dataset_kwargs = dict(
+        ds_kw = dict(
             input_key=input_field,
             target_keys=target_fields,
             max_length=max_length,
             data_pipeline_state=DataPipelineState(),
+            transform_kwargs=transform_kwargs,
+            input_transforms_registry=cls.input_transforms_registry,
         )
 
         return cls(
-            TextClassificationCSVInput(RunningStage.TRAINING, train_file, **dataset_kwargs),
-            TextClassificationCSVInput(RunningStage.VALIDATING, val_file, **dataset_kwargs),
-            TextClassificationCSVInput(RunningStage.TESTING, test_file, **dataset_kwargs),
-            TextClassificationCSVInput(RunningStage.PREDICTING, predict_file, **dataset_kwargs),
-            input_transform=cls.input_transform_cls(
-                train_transform,
-                val_transform,
-                test_transform,
-                predict_transform,
-                max_length=max_length,
-            ),
+            input_cls(RunningStage.TRAINING, train_file, transform=train_transform, **ds_kw),
+            input_cls(RunningStage.VALIDATING, val_file, transform=val_transform, **ds_kw),
+            input_cls(RunningStage.TESTING, test_file, transform=test_transform, **ds_kw),
+            input_cls(RunningStage.PREDICTING, predict_file, transform=predict_transform, **ds_kw),
             **data_module_kwargs,
         )
 
@@ -273,10 +111,12 @@ class TextClassificationData(DataModule):
         val_file: Optional[PATH_TYPE] = None,
         test_file: Optional[PATH_TYPE] = None,
         predict_file: Optional[PATH_TYPE] = None,
-        train_transform: Optional[Dict[str, Callable]] = None,
-        val_transform: Optional[Dict[str, Callable]] = None,
-        test_transform: Optional[Dict[str, Callable]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
+        train_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        val_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        test_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        predict_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        input_cls: Type[Input] = TextClassificationJSONInput,
+        transform_kwargs: Optional[Dict] = None,
         field: Optional[str] = None,
         max_length: int = 128,
         **data_module_kwargs: Any,
@@ -306,26 +146,21 @@ class TextClassificationData(DataModule):
             The constructed data module.
         """
 
-        dataset_kwargs = dict(
+        ds_kw = dict(
             input_key=input_field,
             target_keys=target_fields,
             field=field,
             max_length=max_length,
             data_pipeline_state=DataPipelineState(),
+            transform_kwargs=transform_kwargs,
+            input_transforms_registry=cls.input_transforms_registry,
         )
 
         return cls(
-            TextClassificationJSONInput(RunningStage.TRAINING, train_file, **dataset_kwargs),
-            TextClassificationJSONInput(RunningStage.VALIDATING, val_file, **dataset_kwargs),
-            TextClassificationJSONInput(RunningStage.TESTING, test_file, **dataset_kwargs),
-            TextClassificationJSONInput(RunningStage.PREDICTING, predict_file, **dataset_kwargs),
-            input_transform=cls.input_transform_cls(
-                train_transform,
-                val_transform,
-                test_transform,
-                predict_transform,
-                max_length=max_length,
-            ),
+            input_cls(RunningStage.TRAINING, train_file, transform=train_transform, **ds_kw),
+            input_cls(RunningStage.VALIDATING, val_file, transform=val_transform, **ds_kw),
+            input_cls(RunningStage.TESTING, test_file, transform=test_transform, **ds_kw),
+            input_cls(RunningStage.PREDICTING, predict_file, transform=predict_transform, **ds_kw),
             **data_module_kwargs,
         )
 
@@ -338,10 +173,12 @@ class TextClassificationData(DataModule):
         val_file: Optional[PATH_TYPE] = None,
         test_file: Optional[PATH_TYPE] = None,
         predict_file: Optional[PATH_TYPE] = None,
-        train_transform: Optional[Dict[str, Callable]] = None,
-        val_transform: Optional[Dict[str, Callable]] = None,
-        test_transform: Optional[Dict[str, Callable]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
+        train_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        val_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        test_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        predict_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        input_cls: Type[Input] = TextClassificationParquetInput,
+        transform_kwargs: Optional[Dict] = None,
         max_length: int = 128,
         **data_module_kwargs: Any,
     ) -> "TextClassificationData":
@@ -369,25 +206,20 @@ class TextClassificationData(DataModule):
             The constructed data module.
         """
 
-        dataset_kwargs = dict(
+        ds_kw = dict(
             input_key=input_field,
             target_keys=target_fields,
             max_length=max_length,
             data_pipeline_state=DataPipelineState(),
+            transform_kwargs=transform_kwargs,
+            input_transforms_registry=cls.input_transforms_registry,
         )
 
         return cls(
-            TextClassificationParquetInput(RunningStage.TRAINING, train_file, **dataset_kwargs),
-            TextClassificationParquetInput(RunningStage.VALIDATING, val_file, **dataset_kwargs),
-            TextClassificationParquetInput(RunningStage.TESTING, test_file, **dataset_kwargs),
-            TextClassificationParquetInput(RunningStage.PREDICTING, predict_file, **dataset_kwargs),
-            input_transform=cls.input_transform_cls(
-                train_transform,
-                val_transform,
-                test_transform,
-                predict_transform,
-                max_length=max_length,
-            ),
+            input_cls(RunningStage.TRAINING, train_file, transform=train_transform, **ds_kw),
+            input_cls(RunningStage.VALIDATING, val_file, transform=val_transform, **ds_kw),
+            input_cls(RunningStage.TESTING, test_file, transform=test_transform, **ds_kw),
+            input_cls(RunningStage.PREDICTING, predict_file, transform=predict_transform, **ds_kw),
             **data_module_kwargs,
         )
 
@@ -400,10 +232,12 @@ class TextClassificationData(DataModule):
         val_hf_dataset: Optional[Dataset] = None,
         test_hf_dataset: Optional[Dataset] = None,
         predict_hf_dataset: Optional[Dataset] = None,
-        train_transform: Optional[Union[Callable, List, Dict[str, Callable]]] = None,
-        val_transform: Optional[Union[Callable, List, Dict[str, Callable]]] = None,
-        test_transform: Optional[Union[Callable, List, Dict[str, Callable]]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
+        train_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        val_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        test_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        predict_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        input_cls: Type[Input] = TextClassificationInput,
+        transform_kwargs: Optional[Dict] = None,
         max_length: int = 128,
         **data_module_kwargs: Any,
     ) -> "TextClassificationData":
@@ -431,25 +265,20 @@ class TextClassificationData(DataModule):
             The constructed data module.
         """
 
-        dataset_kwargs = dict(
+        ds_kw = dict(
             input_key=input_field,
             target_keys=target_fields,
             max_length=max_length,
             data_pipeline_state=DataPipelineState(),
+            transform_kwargs=transform_kwargs,
+            input_transforms_registry=cls.input_transforms_registry,
         )
 
         return cls(
-            TextClassificationInput(RunningStage.TRAINING, train_hf_dataset, **dataset_kwargs),
-            TextClassificationInput(RunningStage.VALIDATING, val_hf_dataset, **dataset_kwargs),
-            TextClassificationInput(RunningStage.TESTING, test_hf_dataset, **dataset_kwargs),
-            TextClassificationInput(RunningStage.PREDICTING, predict_hf_dataset, **dataset_kwargs),
-            input_transform=cls.input_transform_cls(
-                train_transform,
-                val_transform,
-                test_transform,
-                predict_transform,
-                max_length=max_length,
-            ),
+            input_cls(RunningStage.TRAINING, train_hf_dataset, transform=train_transform, **ds_kw),
+            input_cls(RunningStage.VALIDATING, val_hf_dataset, transform=val_transform, **ds_kw),
+            input_cls(RunningStage.TESTING, test_hf_dataset, transform=test_transform, **ds_kw),
+            input_cls(RunningStage.PREDICTING, predict_hf_dataset, transform=predict_transform, **ds_kw),
             **data_module_kwargs,
         )
 
@@ -462,10 +291,12 @@ class TextClassificationData(DataModule):
         val_data_frame: Optional[DataFrame] = None,
         test_data_frame: Optional[DataFrame] = None,
         predict_data_frame: Optional[DataFrame] = None,
-        train_transform: Optional[Union[Callable, List, Dict[str, Callable]]] = None,
-        val_transform: Optional[Union[Callable, List, Dict[str, Callable]]] = None,
-        test_transform: Optional[Union[Callable, List, Dict[str, Callable]]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
+        train_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        val_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        test_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        predict_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        input_cls: Type[Input] = TextClassificationDataFrameInput,
+        transform_kwargs: Optional[Dict] = None,
         max_length: int = 128,
         **data_module_kwargs: Any,
     ) -> "TextClassificationData":
@@ -493,25 +324,20 @@ class TextClassificationData(DataModule):
             The constructed data module.
         """
 
-        dataset_kwargs = dict(
+        ds_kw = dict(
             input_key=input_field,
             target_keys=target_fields,
             max_length=max_length,
             data_pipeline_state=DataPipelineState(),
+            transform_kwargs=transform_kwargs,
+            input_transforms_registry=cls.input_transforms_registry,
         )
 
         return cls(
-            TextClassificationDataFrameInput(RunningStage.TRAINING, train_data_frame, **dataset_kwargs),
-            TextClassificationDataFrameInput(RunningStage.VALIDATING, val_data_frame, **dataset_kwargs),
-            TextClassificationDataFrameInput(RunningStage.TESTING, test_data_frame, **dataset_kwargs),
-            TextClassificationDataFrameInput(RunningStage.PREDICTING, predict_data_frame, **dataset_kwargs),
-            input_transform=cls.input_transform_cls(
-                train_transform,
-                val_transform,
-                test_transform,
-                predict_transform,
-                max_length=max_length,
-            ),
+            input_cls(RunningStage.TRAINING, train_data_frame, transform=train_transform, **ds_kw),
+            input_cls(RunningStage.VALIDATING, val_data_frame, transform=val_transform, **ds_kw),
+            input_cls(RunningStage.TESTING, test_data_frame, transform=test_transform, **ds_kw),
+            input_cls(RunningStage.PREDICTING, predict_data_frame, transform=predict_transform, **ds_kw),
             **data_module_kwargs,
         )
 
@@ -525,10 +351,12 @@ class TextClassificationData(DataModule):
         test_data: Optional[List[str]] = None,
         test_targets: Optional[Union[List[Any], List[List[Any]]]] = None,
         predict_data: Optional[List[str]] = None,
-        train_transform: Optional[Union[Callable, List, Dict[str, Callable]]] = None,
-        val_transform: Optional[Union[Callable, List, Dict[str, Callable]]] = None,
-        test_transform: Optional[Union[Callable, List, Dict[str, Callable]]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
+        train_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        val_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        test_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        predict_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        input_cls: Type[Input] = TextClassificationListInput,
+        transform_kwargs: Optional[Dict] = None,
         max_length: int = 128,
         **data_module_kwargs: Any,
     ) -> "TextClassificationData":
@@ -560,20 +388,18 @@ class TextClassificationData(DataModule):
             The constructed data module.
         """
 
-        dataset_kwargs = dict(max_length=max_length, data_pipeline_state=DataPipelineState())
+        ds_kw = dict(
+            max_length=max_length,
+            data_pipeline_state=DataPipelineState(),
+            transform_kwargs=transform_kwargs,
+            input_transforms_registry=cls.input_transforms_registry,
+        )
 
         return cls(
-            TextClassificationListInput(RunningStage.TRAINING, train_data, train_targets, **dataset_kwargs),
-            TextClassificationListInput(RunningStage.VALIDATING, val_data, val_targets, **dataset_kwargs),
-            TextClassificationListInput(RunningStage.TESTING, test_data, test_targets, **dataset_kwargs),
-            TextClassificationListInput(RunningStage.PREDICTING, predict_data, **dataset_kwargs),
-            input_transform=cls.input_transform_cls(
-                train_transform,
-                val_transform,
-                test_transform,
-                predict_transform,
-                max_length=max_length,
-            ),
+            input_cls(RunningStage.TRAINING, train_data, train_targets, transform=train_transform, **ds_kw),
+            input_cls(RunningStage.VALIDATING, val_data, val_targets, transform=val_transform, **ds_kw),
+            input_cls(RunningStage.TESTING, test_data, test_targets, transform=test_transform, **ds_kw),
+            input_cls(RunningStage.PREDICTING, predict_data, transform=predict_transform, **ds_kw),
             **data_module_kwargs,
         )
 
@@ -590,10 +416,12 @@ class TextClassificationData(DataModule):
         val_data_folder: str = None,
         test_data_folder: str = None,
         predict_data_folder: str = None,
-        train_transform: Optional[Dict[str, Callable]] = None,
-        val_transform: Optional[Dict[str, Callable]] = None,
-        test_transform: Optional[Dict[str, Callable]] = None,
-        predict_transform: Optional[Dict[str, Callable]] = None,
+        train_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        val_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        test_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        predict_transform: Optional[Dict[str, Callable]] = TransformersInputTransform,
+        input_cls: Type[Input] = LabelStudioTextClassificationInput,
+        transform_kwargs: Optional[Dict] = None,
         val_split: Optional[float] = None,
         multi_label: Optional[bool] = False,
         max_length: int = 128,
@@ -650,19 +478,17 @@ class TextClassificationData(DataModule):
             multi_label=multi_label,
         )
 
-        dataset_kwargs = dict(data_pipeline_state=DataPipelineState(), max_length=max_length)
+        ds_kw = dict(
+            max_length=max_length,
+            data_pipeline_state=DataPipelineState(),
+            transform_kwargs=transform_kwargs,
+            input_transforms_registry=cls.input_transforms_registry,
+        )
 
         return cls(
-            LabelStudioTextClassificationInput(RunningStage.TRAINING, train_data, **dataset_kwargs),
-            LabelStudioTextClassificationInput(RunningStage.VALIDATING, val_data, **dataset_kwargs),
-            LabelStudioTextClassificationInput(RunningStage.TESTING, test_data, **dataset_kwargs),
-            LabelStudioTextClassificationInput(RunningStage.PREDICTING, predict_data, **dataset_kwargs),
-            input_transform=cls.input_transform_cls(
-                train_transform,
-                val_transform,
-                test_transform,
-                predict_transform,
-                max_length=max_length,
-            ),
+            input_cls(RunningStage.TRAINING, train_data, transform=train_transform, **ds_kw),
+            input_cls(RunningStage.VALIDATING, val_data, transform=val_transform, **ds_kw),
+            input_cls(RunningStage.TESTING, test_data, transform=test_transform, **ds_kw),
+            input_cls(RunningStage.PREDICTING, predict_data, transform=predict_transform, **ds_kw),
             **data_module_kwargs,
         )
